refactor: Lotto 클래스 위치 복구
