+++ src/Lotto.js
@@ -1,5 +1,5 @@
-import { Printer } from '../io/index.js';
-import { LottoValidator } from '../services/index.js';
+import { Printer } from './io/index.js';
+import { LottoValidator } from './services/index.js';
 
 class Lotto {
   #numbers;
