+++ src/models/index.js
@@ -1,4 +1,4 @@
-import Lotto from './Lotto.js';
+import Lotto from '../Lotto.js';
 import Prize from './Prize.js';
 
 export { Lotto, Prize };
